refactor(casino): change types from ulong to long since postgre doesn't support ulong

+++ DiscordBot/Domain/Casino/Game.cs
@@ -56,7 +56,7 @@ public interface ICasinoGame
     public IReadOnlyList<(GamePlayer player, long payout)> EndGame();
 
     public abstract GamePlayerResult GetPlayerGameResult(GamePlayer player);
-    public abstract long CalculatePayout(GamePlayer player, ulong totalPot);
+    public abstract long CalculatePayout(GamePlayer player, long totalPot);
 
     public void Reset();
 
@@ -149,8 +149,8 @@ public abstract class ACasinoGame<TPlayerData, TPlayerAction> : ICasinoGame
     // Default implementation does nothing, override in specific games if needed
     protected virtual void FinalizeGame(List<GamePlayer> players) { }
     public abstract GamePlayerResult GetPlayerGameResult(GamePlayer player);
-    protected ulong GetTotalPot => (ulong)Players.Sum(p => (long)p.Bet);
-    public abstract long CalculatePayout(GamePlayer player, ulong totalPot);
+    protected long GetTotalPot => Players.Sum(p => p.Bet);
+    public abstract long CalculatePayout(GamePlayer player, long totalPot);
 
     /// <summary>
     /// Determines if the game should enter the FINISHED state. <br />
